new: score, lives, explosions; ship changed
still need to fix the shots.. notes on that:
for shots, use the same style as we did with explosions - store a
"centre" point (ship loc at firing time), a heading (player_theta at
firing time), and a "lifeTime".
we increment lifeTime every tick
we draw the shot by using the centre point plus sin/cos of heading times
the lifeTime (or some relative value to it) so it gets furhter away as
it moves... and it can go across the edges.