shot movement, keyPresses
main stuff  i changed is the shot movement, they now go in a straight
line.
Also did minor changes:
1) player is invincible for 2 sec after spawn, and asteroids do not
spawn on top of player (at first.. no promises for when player respawns)
2) ship will no longer be able to fire or be moved if player is
respawning
3) tried to improve handling by increase repetition rate, had to slow
down fire rate using a counter (there's now a delay of several ticks
between shots, although keypresses are registered at a faster rate)
4) changed ship rotation so that the angle changes in the correct
direction of keypresses (left arrow =rotate counter clockwise)

Issues:
1) does pygame have trouble with multiple keypresses? am I handling them
wrong? it seems that ship cannot rotate if its accelerating, and vice
versa,etc...

2) Need to fix the asteroid bounding.. i should be able to calculate the
max distance between any vertex in an asteroid and its centre point, and
use that as the radius of a bounding circle... but because I move the
centrepoint but not any of the other vertices, this doesn't currently
work. I've disabled it for now, but need to fix that.
3) ship movement is still a bit wonky, moving forward is fine but
rotating is kinda wierd.