Added SOUND! and a menu!
menu is a bit iffy still, need to work on it and make it mouse
interactive in addition to keyboard shortcuts, but it looks alright.

sound is annoying as heck! push m to toggle it off.

discovered pygame.Rect objects, am probably going to convert to those
not too long from now.