Fixed thrusters animation
found an issue that seems to be concurrency related, where fetching an
index from the list of asteroids can result in out of bounds error, but
its difficult to reproduce. I put in a try-except clause to get more
info about it. This would be easier in a proper IDE but IDLE doesn't
seem to support debugging.